# Link Generator perf improvements. 