--- a/.gitignore
+++ b/.gitignore
@@ -40,3 +40,4 @@ node_modules
 *launchSettings.json
 global.json
 BenchmarkDotNet.Artifacts/
+*.etl.zip