… src/Microsoft.AspNetCore.Routing/Template/TemplateBinder.cs