@rynowak 2018-09-08
**Adding RVD.TryAdd**

|                           Method |       Mean |     Error |    StdDev |          Op/s |  Gen 0 | Allocated |
|--------------------------------- |-----------:|----------:|----------:|--------------:|-------:|----------:|
|                    AddSingleItem |  27.157 ns | 0.6063 ns | 0.7884 ns |  36,822,777.6 | 0.0015 |     128 B |
|                    AddThreeItems |  52.742 ns | 1.0700 ns | 1.0988 ns |  18,960,350.6 | 0.0015 |     128 B |
|    ConditionalAdd_ContainsKeyAdd |  44.228 ns | 0.9189 ns | 1.2881 ns |  22,610,045.1 |      - |       0 B |
|            ConditionalAdd_TryAdd |  31.000 ns | 0.3510 ns | 0.3111 ns |  32,257,925.3 |      - |       0 B |
|          ForEachThreeItems_Array |  29.312 ns | 0.1200 ns | 0.0937 ns |  34,115,287.1 |      - |       0 B |
|     ForEachThreeItems_Properties |  86.397 ns | 1.5924 ns | 1.4895 ns |  11,574,424.1 |      - |       0 B |
|              GetThreeItems_Array |  41.358 ns | 0.8785 ns | 1.2877 ns |  24,178,907.6 |      - |       0 B |
|         GetThreeItems_Properties | 116.323 ns | 2.3833 ns | 3.0989 ns |   8,596,739.5 |      - |       0 B |
|                    SetSingleItem |  27.056 ns | 0.5854 ns | 0.6012 ns |  36,960,761.4 | 0.0014 |     128 B |
|                  SetExistingItem |   8.573 ns | 0.2228 ns | 0.1739 ns | 116,643,503.0 |      - |       0 B |
|                    SetThreeItems |  50.656 ns | 1.1139 ns | 3.2138 ns |  19,740,867.8 | 0.0014 |     128 B |
|      TryGetValueThreeItems_Array |  43.504 ns | 0.9272 ns | 1.8302 ns |  22,986,549.2 |      - |       0 B |
| TryGetValueThreeItems_Properties | 107.692 ns | 4.1687 ns | 4.9625 ns |   9,285,738.1 |      - |       0 B |

Notice that the new TryAdd is significantly faster than the old pattern.