@rynowak 2018-09-08
**Use registers for RVD array accesses**

|                           Method |       Mean |     Error |    StdDev |     Median |          Op/s |  Gen 0 | Allocated |
|--------------------------------- |-----------:|----------:|----------:|-----------:|--------------:|-------:|----------:|
|                    AddSingleItem |  26.665 ns | 0.6040 ns | 1.1638 ns |  26.457 ns |  37,501,751.5 | 0.0014 |     128 B |
|                    AddThreeItems |  52.663 ns | 1.1245 ns | 1.3810 ns |  52.048 ns |  18,988,676.1 | 0.0014 |     128 B |
|    ConditionalAdd_ContainsKeyAdd |  40.976 ns | 0.8786 ns | 1.8145 ns |  40.109 ns |  24,404,551.5 |      - |       0 B |
|            ConditionalAdd_TryAdd |  28.928 ns | 0.6439 ns | 1.0398 ns |  28.967 ns |  34,568,660.7 |      - |       0 B |
|          ForEachThreeItems_Array |  28.222 ns | 0.6237 ns | 1.1086 ns |  28.173 ns |  35,433,723.5 |      - |       0 B |
|     ForEachThreeItems_Properties |  86.009 ns | 1.7731 ns | 2.9133 ns |  84.871 ns |  11,626,688.6 |      - |       0 B |
|              GetThreeItems_Array |  39.507 ns | 0.5708 ns | 0.5339 ns |  39.641 ns |  25,311,751.7 |      - |       0 B |
|         GetThreeItems_Properties | 110.002 ns | 1.2838 ns | 1.1380 ns | 110.307 ns |   9,090,736.0 |      - |       0 B |
|                    SetSingleItem |  26.825 ns | 0.7262 ns | 0.9184 ns |  26.603 ns |  37,278,651.0 | 0.0014 |     128 B |
|                  SetExistingItem |   8.049 ns | 0.2207 ns | 0.2453 ns |   7.948 ns | 124,244,539.7 |      - |       0 B |
|                    SetThreeItems |  50.941 ns | 1.1584 ns | 1.9982 ns |  50.245 ns |  19,630,564.1 | 0.0014 |     128 B |
|      TryGetValueThreeItems_Array |  39.181 ns | 0.4336 ns | 0.3136 ns |  39.250 ns |  25,522,890.6 |      - |       0 B |
| TryGetValueThreeItems_Properties | 112.249 ns | 2.0146 ns | 1.8845 ns | 112.823 ns |   8,908,765.0 |      - |       0 B |