@rynowak 2018-09-08
**Inlining for RVD TryGetValue**

|          Method |       Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|---------------- |-----------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
|        Baseline |   339.8 ns |  2.371 ns |  2.218 ns | 2,942,933.5 |   1.00 |     0.00 | 0.0010 |     112 B |
|      TreeRouter | 2,125.0 ns |  8.415 ns |  6.085 ns |   470,594.7 |   6.25 |     0.04 | 0.0114 |    1168 B |
| EndpointRouting | 3,581.5 ns | 69.821 ns | 77.606 ns |   279,211.2 |  10.54 |     0.23 | 0.0153 |    1720 B |

This actually makes our baseline much faster.