@rynowak 2018-09-12
## Rewrite of TemplateBinder.GetValues

|          Method |       Mean |     Error |   StdDev |     Median |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|---------------- |-----------:|----------:|---------:|-----------:|------------:|-------:|---------:|-------:|----------:|
|        Baseline |   381.6 ns |  7.588 ns | 18.47 ns |   374.4 ns | 2,620,389.3 |   1.00 |     0.00 | 0.0010 |     112 B |
|      TreeRouter | 2,146.3 ns | 42.330 ns | 92.02 ns | 2,135.3 ns |   465,926.8 |   5.64 |     0.35 | 0.0114 |    1136 B |
| EndpointRouting | 3,566.9 ns | 54.912 ns | 53.93 ns | 3,553.9 ns |   280,353.0 |   9.37 |     0.44 | 0.0153 |    1600 B |

TreeRouter getting much faster with these changes, as I'm improving the shared stuff.