@rynowak 2018-09-12
@JamesNK - I plan to sort of keep iterating on this for the next day or so before merging any of this stuff. I'll probably fork the RVD changes into a separate PR since that's pretty isolated.