@rynowak 2018-09-12
@JamesNK - OK I think I've shaken out most of the obvious stuff here. Could you take another look?